Added gauss.js, it will contain the gauss elimination implementation, which applies gauss elimination on arbitrary matrices lu.js will contain LU decomposition, which is the standard way of decomposing a matrix for calculating its inverse, determinant, rank and other stuff. compmat is just a sketch on finding the matrix representation for complex numbers (x+iy represented as a 2x2 matrix). updates on matrix.js, added whos, improved size, fixed add_s, now all functions uses operator_type as format. not the other way around.
added rand, not finished yet.

Added some comments.

One problem is that premature optimisation affects the code, I did build all matrix code using arrays of,[... of], arrays for simplicity and also allowed
them to be sparse. This might not work out so well in terms of efficency regarding sparseness so maybe I'll toss the support for sparse matrices but keep
 the multidim arrays for simplicity, we'll see.