Added getrow and setrow on matrix, also added iterator firstrow, next element. It makes it much easier to traverse an entire matrix and process each element. The support for sparse matrices still hurts in all areas, but I fixed the add_m code and gauss/gauss_jordan now handles sparse matrices by inserting zeros in empty rows, not very efficient and sort of kills the idea of using sparse matrices. The get/set member will instead set and get submatrices and also support flexible indexing such as get(1, 2, :) etc (using special syntax for ':' operator)
gauss/gauss_jordan now works for real matrices, ivnerse was implemented using gauss/jordan on identity augmented matrix.
I have to do some verification on the fact that the inverse can be achieved using gauss/jordan with pivoting but without
the use of the pivot array.

Maybe that is expected and that the pivot array is only needed when solving linear systems such as B = A\x.

The Next Big Thing will be support for complex matrices.